two more standalone solutions for problem 1

+++ src/main/kotlin/standalone/p0001s1/standalone1_4.kt
@@ -1,8 +1,8 @@
 package standalone.p0001s1
 
-val end = 1000
+val end = 1000L
 
-fun sumOfMultiples(num: Int) = (num until end step num).sum()
+fun sumOfMultiples(num: Long) = (num until end step num).sum()
 
 fun main() {
     println(sumOfMultiples(3) + sumOfMultiples(5) - sumOfMultiples(15))
